Rename package name from credential to credentials
