+++ quarkus/service/src/main/java/org/apache/polaris/service/quarkus/config/QuarkusProducers.java
@@ -68,19 +68,17 @@ import org.apache.polaris.service.context.RealmContextConfiguration;
 import org.apache.polaris.service.context.RealmContextResolver;
 import org.apache.polaris.service.events.PolarisEventListener;
 import org.apache.polaris.service.identity.ServiceIdentityConfiguration;
-import org.apache.polaris.service.identity.mutation.EntityMutationConfiguration;
 import org.apache.polaris.service.quarkus.auth.QuarkusAuthenticationConfiguration;
 import org.apache.polaris.service.quarkus.auth.QuarkusAuthenticationRealmConfiguration;
 import org.apache.polaris.service.quarkus.auth.external.tenant.OidcTenantResolver;
 import org.apache.polaris.service.quarkus.catalog.io.QuarkusFileIOConfiguration;
 import org.apache.polaris.service.quarkus.context.QuarkusRealmContextConfiguration;
 import org.apache.polaris.service.quarkus.context.RealmContextFilter;
-import org.apache.polaris.service.quarkus.credential.QuarkusPolarisCredentialManagerConfiguration;
+import org.apache.polaris.service.quarkus.credentials.QuarkusPolarisCredentialManagerConfiguration;
 import org.apache.polaris.service.quarkus.events.QuarkusPolarisEventListenerConfiguration;
 import org.apache.polaris.service.quarkus.identity.QuarkusRealmServiceIdentityConfiguration;
 import org.apache.polaris.service.quarkus.identity.QuarkusServiceIdentityConfiguration;
 import org.apache.polaris.service.quarkus.identity.QuarkusServiceIdentityRegistryConfiguration;
-import org.apache.polaris.service.quarkus.identity.mutation.QuarkusEntityMutationConfiguration;
 import org.apache.polaris.service.quarkus.persistence.QuarkusPersistenceConfiguration;
 import org.apache.polaris.service.quarkus.ratelimiter.QuarkusRateLimiterFilterConfiguration;
 import org.apache.polaris.service.quarkus.ratelimiter.QuarkusTokenBucketConfiguration;
@@ -199,12 +197,6 @@ public class QuarkusProducers {
       @Any Instance<ServiceIdentityRegistryFactory> serviceIdentityRegistryFactories) {
     return serviceIdentityRegistryFactories.select(Identifier.Literal.of(config.type())).get();
   }
-
-//  @Produces
-//  public EntityMutationConfiguration entityMutationConfiguration(
-//      QuarkusEntityMutationConfiguration config) {
-//    return config;
-//  }
 
   @Produces
   public PolarisCredentialManagerFactory credentialManagerFactory(
